rubber-grading-backend/backend/requirements.txt
flask
pymongo
python-dotenv
flask-cors
tensorflow
numpy
Pillow
opencv-python-headless